refactor(graduate): 논문 및 자격증 제출 스타일 통일
