+++ apps/graduate/src/pages/client/certification/ui/CertificationPage.tsx
@@ -13,8 +13,7 @@ import type { CSSProperties } from 'react';
 import { Section } from '~/shared/components';
 import { ROUTE } from '~/shared/constants';
 
-import * as styles from '~/pages/client/thesis/styles/ThesisPage.css';
-import { vars } from '~/vars.css';
+import * as styles from '~/shared/styles/SubmissionPage.css';
 
 export default function CertificationPage() {
   const FileUpload = () => {
@@ -43,13 +42,6 @@ export default function CertificationPage() {
     },
   ];
 
-  const panelStyle: React.CSSProperties = {
-    marginBottom: vars.spacing.md,
-    background: vars.colors.white,
-    borderRadius: vars.radius.lg,
-    border: 'none',
-  };
-
   const { Dragger } = Upload;
 
   const props: UploadProps = {
@@ -73,43 +65,24 @@ export default function CertificationPage() {
   };
 
   return (
-    <div
-      style={{
-        display: 'grid',
-        placeItems: 'center',
-        width: '100%',
-        maxWidth: '768px',
-        margin: 'auto',
-        gap: vars.spacing.xl,
-      }}
-    >
+    <div className={styles.container}>
       <Section.Header subtitle='자격증 증빙 서류를 제출해주세요.'>
         자격증 증빙 서류 제출
       </Section.Header>
       <Collapse
         size='large'
         bordered={false}
-        items={items(panelStyle)}
+        items={items(styles.panelStyle)}
         defaultActiveKey={['certification']}
         className={styles.collapse}
       />
-      <div style={{ display: 'flex', gap: vars.spacing.md, width: '100%' }}>
-        <Link
-          to={ROUTE.HOME}
-          style={{
-            width: '100%',
-          }}
-        >
+      <div className={styles.buttonContainer}>
+        <Link to={ROUTE.HOME} className={styles.buttonWrapper}>
           <Button size='large' type='primary' className={styles.button}>
             이전으로
           </Button>
         </Link>
-        <Link
-          to={ROUTE.HOME}
-          style={{
-            width: '100%',
-          }}
-        >
+        <Link to={ROUTE.HOME} className={styles.buttonWrapper}>
           <Button size='large' className={styles.button} type='primary'>
             제출하기
           </Button>
