--- apps/graduate/src/pages/client/thesis/styles/ThesisPage.css.ts
@@ -1,14 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-import { vars } from '~/vars.css';
-
-export const collapse = style({
-  width: '100%',
-  backgroundColor: 'transparent',
-});
-
-export const button = style({
-  width: '100%',
-  padding: vars.spacing.lg,
-  borderRadius: vars.radius.lg,
-});
